fix(mobile): resolve M11.3 CI failures on iOS and Android
Move instructor insights API methods out of LMSAPIFeatures to satisfy
SwiftLint file length, and align Android screens with existing Compose
patterns for auth token, offline state, and empty states.
